final changes

+++ server-Chat-App/src/test/java/org/server/chatapp/dao/implement/MessageDaoImplTest.java
@@ -9,7 +9,6 @@ import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-import static org.junit.jupiter.api.Assertions.*;
 
 @TestMethodOrder(MethodOrderer.OrderAnnotation.class)
 class MessageDaoImplTest {
@@ -42,7 +41,7 @@ class MessageDaoImplTest {
         testMessage.setSentAt(LocalDateTime.now());
         testMessage.setDeleted(false);
 
-        int inserted = messageDao.insert(testMessage);
+        long inserted = messageDao.insert(testMessage);
         assertEquals(1, inserted, "Insert should affect 1 row");
         assertTrue(testMessage.getId() > 0, "Inserted message should have generated ID");
     }
